Address thermo review findings
